feat(nvim): migrate from lazy.nvim to declarative home-manager nixvim
Replaces the lazy.nvim setup with nixvim on khazad-dum (opt-in via
agindin.neovim.enable). Treesitter is now declarative on the main branch,
fixing the Neovim 0.12 startup crash caused by the archived nvim-treesitter
master branch. Plugins are pinned via flake.lock and the build is hermetic
(wrapRc), so stale lazy state in ~/.config or ~/.local/share can no longer
leak in.

Custom Lua (nav remaps, catppuccin, lualine, bufferline, LSP, which-key)
lives in common/nvim/config.lua; everything else is nixvim modules. Drops
opencode.nvim and telescope-tabs.
