+++ hosts/khazad-dum/home.nix
@@ -8,6 +8,7 @@
     claude-code.enable = true;
     claude.desktop.enable = true;
     lsp.enable = true;
+    neovim.enable = true;
 
     mcp = {
       enable = true;
